+++ src/main/java/spoon/experimental/modelobs/ActionBasedChangeListener.java
@@ -16,6 +16,12 @@
  */
 package spoon.experimental.modelobs;
 
+import spoon.experimental.modelobs.action.Action;
+import spoon.experimental.modelobs.action.AddAction;
+import spoon.experimental.modelobs.action.DeleteAction;
+import spoon.experimental.modelobs.action.DeleteAllAction;
+import spoon.experimental.modelobs.action.UpdateAction;
+
 /**
  * notifies all change on the AST
  */
