+++ src/main/java/spoon/experimental/modelobs/FineModelChangeHandler.java
@@ -16,6 +16,11 @@
  */
 package spoon.experimental.modelobs;
 
+import spoon.experimental.modelobs.action.Action;
+import spoon.experimental.modelobs.action.AddAction;
+import spoon.experimental.modelobs.action.DeleteAction;
+import spoon.experimental.modelobs.action.DeleteAllAction;
+import spoon.experimental.modelobs.action.UpdateAction;
 import spoon.experimental.modelobs.context.ListContext;
 import spoon.experimental.modelobs.context.MapContext;
 import spoon.experimental.modelobs.context.ObjectContext;
