+++ src/main/java/spoon/experimental/modelobs/action/DeleteAction.java
@@ -14,28 +14,32 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-package spoon.experimental.modelobs;
+package spoon.experimental.modelobs.action;
 
 import spoon.experimental.modelobs.context.Context;
 
-public class AddAction<T>  extends Action {
-	private T newValue;
+/**
+ * defines the delete action
+ * @param <T>
+ */
+public class DeleteAction<T> extends Action {
+	private T oldValue;
 
-	public AddAction(Context context, T newElement) {
+	public DeleteAction(Context context, T oldValue) {
 		super(context);
-		this.newValue = newElement;
+		this.oldValue = oldValue;
 	}
 
 	@Override
 	public T getChangedValue() {
-		return getNewValue();
+		return getRemovedValue();
 	}
 
 	/**
-	 * Returns the added element
-	 * @return the new element
+	 * Returns the removed element
+	 * @return the removed element
 	 */
-	public T getNewValue() {
-		return newValue;
+	public T getRemovedValue() {
+		return oldValue;
 	}
 }
