+++ src/main/java/spoon/experimental/modelobs/action/UpdateAction.java
@@ -14,7 +14,7 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-package spoon.experimental.modelobs;
+package spoon.experimental.modelobs.action;
 
 import spoon.experimental.modelobs.context.Context;
